cast coef_scale as double

+++ src/fame/model/custom_layers.py
@@ -1,3 +1,4 @@
+import numpy as np
 import tensorflow as tf
 from tensorflow.keras import layers
 
@@ -5,7 +6,7 @@ from tensorflow.keras import layers
 class CoefSinhLayer(layers.Layer):
     def __init__(self, coef_scale, **kwargs):
         super(CoefSinhLayer, self).__init__(**kwargs)
-        self.coef_scale = coef_scale
+        self.coef_scale = np.float64(coef_scale)
         
     def call(self, inputs):
         return tf.multiply(self.coef_scale, tf.math.sinh(inputs))
